requirements.txt
paramiko==2.7.2
python-dotenv==0.19.0
six==1.16.0
pre-commit==2.17.0